Add --with-custom-python option to virtualenv formulæ
Installing with `-s` in order that `:python` dependencies do not install
Homebrew's python has some poor ergonomics: it's not discoverable and it
doesn't persist with `brew upgrade`. This adds an option to every
virtualenv-based formula with a discoverable name that has no effect
except to force a source build and which will be memorized for upgrades.